+++ client/src/App.tsx
@@ -1,6 +1,7 @@
 import React from "react";
 import GlobalThemeProvider from "styles/GlobalThemeProvider";
 import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
+import MainPage from "components/templates/MainPage";
 import TestPage from "components/templates/TestPage";
 
 const App: React.FC = (): JSX.Element => {
@@ -8,8 +9,8 @@ const App: React.FC = (): JSX.Element => {
     <GlobalThemeProvider>
       <Router>
         <Switch>
-          <Route exact path="/" component={TestPage} />
-          {/* <Route exact path="/" component={} /> */}
+          <Route exact path="/test" component={TestPage} />
+          <Route exact path="/" component={MainPage} />
           {/* <Route path="/upload" component={} /> */}
           {/* <Route component={NotFoundPage} /> */}
         </Switch>
