+++ client/src/components/templates/MainPage/style.ts
@@ -1,0 +1,14 @@
+import styled from "styled-components";
+
+export const Container = styled.div`
+  margin-top: 30px;
+  display: flex;
+  justify-content: center;
+`;
+
+export const Elements = styled.div`
+  display: block;
+  align-items: center;
+  justify-content: center;
+  width: 343px;
+`;
